Alias Tcflag to UInt64 instead of UInt32
This is probably system dependent but UInt64
should work on more systems until we can parse
header files

+++ src/io/console.cr
@@ -1,6 +1,6 @@
 lib Termios
   alias Cc = Char
-  alias Tcflag = UInt32
+  alias Tcflag = UInt64
 
   struct Struct
     iflag : Tcflag
